+++ src/settings/TemplatedQueryRepos.js
@@ -9,10 +9,10 @@ import { Loading, Row, Col, TextField, IconButton } from '@folio/stripes/compone
 
 
 function TemplatedQueryRepos(props) {
-  const [ConnectedConfigManager, setConfigManager] = useState();
+  const [ConnectedConfigManager, setConnectedConfigManager] = useState();
 
   useEffect(() => {
-    setConfigManager(props.stripes.connect(ConfigManager));
+    setConnectedConfigManager(props.stripes.connect(ConfigManager));
   }, [props.stripes]);
 
   if (!ConnectedConfigManager) return <Loading size="xlarge" />;
@@ -28,6 +28,7 @@ function TemplatedQueryRepos(props) {
 
   const userLabel = <FormattedMessage id="ui-ldp.settings.tqrepos.user" />;
   const repoLabel = <FormattedMessage id="ui-ldp.settings.tqrepos.repo" />;
+  const branchLabel = <FormattedMessage id="ui-ldp.settings.tqrepos.branch" />;
   const dirLabel = <FormattedMessage id="ui-ldp.settings.tqrepos.directory" />;
 
   return (
@@ -43,35 +44,30 @@ function TemplatedQueryRepos(props) {
         <FieldArray name="repos">
           {({ fields }) => (
             <>
-              {fields.map((subname, index) => (
-                <div key={index}>
-                  <Row>
-                    <Col xs={11}>
-                      <Field name={`${subname}.user`} label={userLabel} component={TextField} />
-                      <Field name={`${subname}.repo`} label={repoLabel} component={TextField} />
-                      <Field name={`${subname}.branch`} label={repoLabel} component={TextField} />
-                      <Field name={`${subname}.dir`} label={dirLabel} component={TextField} />
-                    </Col>
-                    <Col xs={1}>
-                      <IconButton icon="trash" onClick={() => fields.remove(index)} />
-                    </Col>
-                    <Col xs={12}>{
-                        // eslint-disable-next-line @calm/react-intl/missing-formatted-message
-                      }
-                      https://github.com/
-                      {fields.value[index].user}
-                      /
-                      {fields.value[index].repo}
-                      /tree/
-                      {fields.value[index].branch}
-                      /
-                      {fields.value[index].dir}
-                      <hr />
-                      <br />
-                    </Col>
-                  </Row>
-                </div>
-              ))}
+              {fields.map((subname, index) => {
+                const val = fields.value[index];
+                const url = `https://github.com/${val.user}/${val.repo}/tree/${val.branch}/${val.dir}`;
+                return (
+                  <div key={index}>
+                    <Row>
+                      <Col xs={11}>
+                        <Field name={`${subname}.user`} label={userLabel} component={TextField} />
+                        <Field name={`${subname}.repo`} label={repoLabel} component={TextField} />
+                        <Field name={`${subname}.branch`} label={branchLabel} component={TextField} />
+                        <Field name={`${subname}.dir`} label={dirLabel} component={TextField} />
+                      </Col>
+                      <Col xs={1}>
+                        <IconButton icon="trash" onClick={() => fields.remove(index)} />
+                      </Col>
+                      <Col xs={12}>
+                        <a target="_blank" rel="noreferrer" href={url}>{url}</a>
+                        <hr />
+                        <br />
+                      </Col>
+                    </Row>
+                  </div>
+                );
+              })}
               <IconButton icon="plus-sign" onClick={() => fields.push('')} />
             </>
           )}